+++ src/main/java/fk/stardust/localizer/sbfl/Wong1.java
@@ -28,7 +28,7 @@ public class Wong1<T> extends AbstractSpectrumBasedFaultLocalizer<T> {
 
     @Override
     public double suspiciousness(final INode<T> node) {
-        return node.getIF();
+        return (double) node.getIF();
     }
 
     @Override
